+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release of the plugin.
+- Fetches list of installed Android applications (name, package, version, icon).
+- Provides option to include/exclude system apps.
+- Includes `InstalledApp` model class.
+- Includes `AppIconWidget` helper widget for displaying icons.